move gui to _gui
