+++ brownie/__init__.py
@@ -3,7 +3,7 @@
 from .project import compile_source, run
 from .network import accounts, alert, history, rpc, web3
 from brownie.network.contract import Contract  # NOQA: F401
-from brownie.gui import Gui
+from brownie._gui import Gui
 from brownie._config import CONFIG as config
 from brownie.convert import Wei
