+++ brownie/_gui/bases.py
