+++ brownie/_gui/console.py
