+++ brownie/_gui/opcodes.py
