+++ brownie/_gui/report.py
