+++ brownie/_gui/root.py
