+++ brownie/_gui/styles.py
